Split react player into separate component, add volume control, fix jump-to-middle bug
